A project Qualifier is only for the parsed project.
That is the Qualifier is the actual string explicitly placed before
the reserved work project on the source file. Then on a project view
we use the term Kind which is more generic and it may or may not be
the same as the qualifier. The project kind is determined from
looking at the project content if the qualifier is not present.

Part of P113-046.

Change-Id: Idc3128412076af86ca3b386b9e7c3a61f0a725ee