Add attribute description knowledge base.
This package enclose descriptions for an attribute. If the index is
allowed or optional, if the value is a single value or a list and
in which context (project's root or specific package) an attribute is
allowed.

Note that in this initial version attribute context is too wide and
would be straightened up in future commits.

Part of P113-046.

Change-Id: I2d512aed9c6a2d1fa750cd262f4781a7646aba97