Introduce No_Case Name/Value types.
Rework Name_Type support.

A Name_Type is now a type and handled without case-sensitivity. An
Optional_Name_Type is introduced which allows empty value.

The code is changed to take this into account. For example there is
no more need to Less_Case_Insensitive or Equal_Case_Insensitive as the
Name_Type's "=" and "<" operator are already case insensitive.

Part of P113-046.

Change-Id: Ie16e5e7add49a16ffffee2610afafdf7474c8c5c