Enhance Naming_Package routine to support configuration project.
Now that a configuration project is supported and that we can access
it from a view we can use the Naming package from the configuration
project if present.

So now Naming_Package returns the naming from the current view,
or the configuration project and if neither are present use the
built-in naming.

Add corresponding regression test.

Part of P113-046 and P728-037.

Change-Id: I029da34ffeb32189e7dbaad2e8e479fed8df5a30