Store the project tree in Views.
Instead of storing the configuration view we directly store the
tree (actually an access to the tree). For this a project tree
object is now limited and is self referenced (Rosen's trick).

The Load routine is now a procedure to comply with the limited
tree object. This permits some clean-up as some routines are not
needed anymore.

Note that the tree will be needed to be able to record error messages
(in the tree's log object) while parsing sources. So this is globally
a clean-up to prepare future work.

The regression tests are adjusted accordingly.

Part of P113-046.

Change-Id: I6f5068a11c2f346d8df8c37b35c6d6b8898e9745