Fix sources key used in the set.
We need to know if two sources are the same and this is not if the
filename is the same. For unit based language we need to check for
a spec or body of a given unit. The same body for example can be
changed via an extending project using a different naming convention.

Part of P113-046.

Change-Id: I12d4aa0597e3cc16abb653c364c223d06c586677