Simplify Naming package circuitry.
First the Dot_Replacement and Is_Standard_GNAT_Naming is computed only
once in the Sources routine for the given view. At the same time the
code is simply as we know that the corresponding attributes are always
present. That is, for a project defining the Naming package, the final
content is the one of the built-in Naming package with all project's
defined Naming's attribute overwriting the default values.

Part of P113-046.

Change-Id: Ica8e32c4be56ecbe6d8bbf78f56029c62822b141