Implement built-ins in a separate package.
This is cleaner and will avoid cluttering the code with complex
built-in implementation.

Note that this also fixes the External semantics where a value is
mandatory if there is no default-value specified. That is, if there
is no corresponding value in the context we record an error message.

Add corresponding regression test.

For this we are now requiring that a context (possibly empty) be
passed when loading the project.

Continued work for P113-046.

Change-Id: Icbf14c6bdb3401ed405a41c242211fda011c9a96