Rework the build procedure to make it simpler.
Import Quex source code. We do not require the developer to install
Quex anymore. The langkit generated sources being already part of the
repository it is now possible to build LibGPR2 with a simple make
from the top directory.

On the directory langkit we just keep a script to regenerate the
grammar from a new version of langkit.

Change-Id: Ib3731f4298a8378f234a35e4b8515bf6cd690738