Fix support for extended project.
The extended projects where not part of the iterator and not properly
hanlded for recursive iterator. This was breaking the parsing as some
project were not properly analysed in order and references were not
found.

Add corresponding regression test.

Also fix expected output of a previous test as now the extended
projects gets reported.

Continued work for OB10-014.

Change-Id: Id75581c1e0730cefee2e66d70804ab12bbc3af4e