Add initial support for the runtime project.
When a configuration file is created we also create a built-in
runtime project. This project can be referenced with Runtime' in
user's project as any other project.

This initial implementation does not cover all cases. The only
simple case taken into account if the Runtime_Dir attribute of
the configuration project.

Add corresponding regression test.

Part of Q614-013.

Change-Id: I89fe964f72d654a43baf870d87292947440799e0