Fix handling of undefined attribute reference.
When an attribute reference is not found we return an empty list
instead of an empty single value. This ensure that we do not create
a list with two items one being an empty string. Also make sure that
the Attr value is defined and code do not depends on unset values.

Should fix the regression tests on x86-linux.

For Q912-010.
Part of OB10-014.

Change-Id: I3f82ebba4d756fdaceddebb220769bc6d73db8a9