Add a way to retrieve an imported project object.
This is based on base-name as it is actually what is found on the
project.

   for Attribute use Project.Attribute.

Here, Project is the basename for an imported project in the list.
The algorithm here is just to look over all items and compare the
base-name. This is time consuming and will need to be reworked.

Part of OB10-014.

Change-Id: I59e40c9a780edce03492c73e1145581e351aa8d3