Add support for setting a project tree given a project view.
This can now be used to associate a configuration project with its
corresponding project tree. This will remove a potential ambiguity
if two project trees are using a configuration project with the same
name. At the same time it makes it possible to clean up the code by
avoiding a need for special case for configuration project.

Part of OB10-014.

Change-Id: I55d9c59e0881e3db690bda74c01fdd6551792d65