Rework code to have better error message for missing project files.
The error message reported for a missing extended or withed project
was not very useful. This patch fixes this. To have a better error
message for extended projects we have also changed the API in the
parser to record the extended project together with its sloc.

Add corresponding regression test.

Part of OB10-014.

Change-Id: I829151774801e15f9e553e500ccd0dc210bd4c46