Change implementation for an import set.
This is used only for stroring the with clause value as found in
the project file. There is no value to keep the base-name and full-name
as actually there is no full-name at all in play here.

So the implementation is now simpler and still allows detection of
duplicate withed unit. This will also make it possible to check
for duplicate project name found in the withed clauses of a project.

Part of R509-005.

Change-Id: I80f11bba9c8f1b8e40c4e0328b9e9be7446c3d20