Add support for the notion of child project.
In Parser.Project.Parse:
When a project is declared as a child project, e.g. "P.Q", store P
as Parent_Name and Q as Name. Also, check that the "end <proj>;" node
refers to the same identifier(s).

In Parser.Project.Process:
Check that a child project imports (in a non-limited way)
or extends its parent.

Change-Id: Ia0a2e237b2c51543a54ef65ab14646b02fcb7114