Attribute's Name and Value/Values are now source referenced.
The Name is a Source_Reference.Identifier.Object and the Value
is a Source_Reference.Value.Object. This is needed to be able
to record the source references for values. The API is now also
cleaner, we were passing the Name (as Name_Type) and the corresponding
sloc to each routine. Now the slock for the Name is into the
corresponding object.

This change also support Name and Value for variables and types.

Adapt the code to this new API.

Change-Id: Ica2bd2ab0f4666c58378d7b8b7748b7b58613269