New Name_Values.Value accessor, and use it in Update_Sources.
This Value function fetches the Source_Reference.Value.Object from a
Value_Type key. The internal V_Set becomes a V_Map to make it fast.

It is used in Handle_File.Check_Naming_Exceptions when checking for
entries in Specification_Exceptions that are duplicated in
Implementation_Exceptions: reporting the corresponding Sloc is easier.

Change-Id: I69185d905b83cd0c584768ab5ab028764975d394