Fix design violation where routine can return an undefined object.
We introduce the necessary routine to check for existence of the
corresponding attributes. The spec aspect for getting the attribute
value are now simpler and require that the attribute be present.
Likewise, the implementation is simpler as we know that the attribute
exists when inside the routine.

Part of S502-021.

Change-Id: Ie8f001b2aa33d84b5978ffc6a1ba89558548f680