Minor improvements
* src/lib/gpr2-parser-project.adb:
(Parse): Use Path_Name.Object method Exists instead of call to
 Ada.Directories. Put Filename.Value into "does not exists" error message.
 Do not show line=1 and column=1 in the error message.

* src/tools/gprls-options-build_from_command_line.adb:
Do not check for project existence because the full path name could be not
known at this stage and will be find out in search paths later.

* src/tools/gprls-process.adb:
Remove exception handlers around Tree.Load_Autoconf and Tree.Update_Sources
because there is exception handler at the end of the procedure.

Change-Id: Ie8a6758eb8cebca870d5aab7be180f2fe21ffd7b