Allow Update_Sources not to stop on missing sources
SC02-017

We can detect all missing source files at once with this modification.
Update_Sources ends the source files processing at the first nonexistent
source file defined in the Source_Files attribute. This is inconvenient
when we need to detect all missing source files, which may be necessary if
we want to generate missing source files for the existing projects tree.

Change-Id: I2504b65bb32360ef59fffb1960c266e2012b34b0