Update Project.Source API for other parts.
A project source now embedded a source. The other-part is handled here
and properly providing interfaces to get the other part of any unit
in the projectt's source. That is, this brings support for other-part
for multi-unit sources.

A Path_Name routine is also provided to get the path-name of the
corresponding source file.

For SC30-005.

Change-Id: I87c89e67f0833966d0979fed56b103f37ebadade