Simplify memory handing and implement function to free answer buffers.
Binding was recompiled with -fstack-usage. Use directly c_chars_ptr
type for C_Request and C_Answer and re-organize the code so that no
dynamic allocation is done on the stack (and thus prevent potential
stack overflows). The only remaining dynamic allocation on the stack
is in fact bounded (in Set_Address).

Add gpr2_free_answer C function.

Part of T526-029

Change-Id: I82af0a2ca828747a34cb8b00f5f6b7b98e85d3e6