GPRLS aggregated processing fixes.
T819-014

Intruduce Namespace_Root project for the each project.
It is root of the projects subtree where source simple file name is unique.
For aggregate library and for none aggregate projects it is the
Tree.Root_Project. For aggregate projects each aggregated project is root of
its subtree.
Add index to find a source by its simple name and name of its subtree root
project name.
Use this index to find source dependencies.

Change-Id: Iee15bbf8557c91b32ba9c96aa9af398bd35121e4