Use naming package from extended project if defined
TC08-004

* src/lib/gpr2-project-definition.adb
(Naming_Package): Avoid double lookup in packages.
 Search for Naming package in extended project if it absent in the extending
 one.
(Update_Sources): Do not check that project is not abstract and not
 config at the second time, it is already tested at the start of routine.
 Check that all naming exceptions is used only in case when Naming package
 is not inherited from extended project.

* src/lib/gpr2-project-tree.adb (Set_Source):
Sources from extending projects has more priority than from extended one to
be in the Tree.Rooted_Sources.

* src/lib/gpr2-project-view.ad[sb] (Is_Extending):
Add Parent parameter to check that the Parent is transitively extended for
the Self.

Change-Id: I8b728e81e2863f2482665e97fb83af5c731332a2