Fix and enhance the Ada language parser.
The separate status is not set after parsing the package spec/body
otherwise the status was rewritten.

The parser is change to actually reset the source-info data only
if something has been parsed otherwise we keep part of the information
coming from another parser or initialized while parsing the project.

With this parser the dependencies are properly known only for the
first unit in a file. All other units are kept as-is, and the
information are exactly what has been induced from the project
naming package being default one or a specific one with multi-unit
information. We cannot guarantee that the file is having those units
though.

Update some expected output to match the new partial parsed information.

no-tn-check

Change-Id: If314761c67a26800de16310594138642249c326b