Use toolchain related attributes for autoconf
Add attributes Required_Toolchain_Version and Toolchain_Path
to the list of recognized attributes.
Use their values and value of Toolchain_Name for auto
configuration, when present.
Improve diagnostic message when toolchain is not found and
path, name and/or version are set via project attributes.
Fix diagnostic message for non-default runtime ("runtime"
was missing).

for U119-039

Change-Id: I8a4afba9b9d1f2b5adca192c77df667465288e04