Fix extends and extends all handling.
Extending a project allows to modify its included projects by variants that
are also extensions of those imports. This means that a new view needs to be
created that takes into account this import substitution. Such substitution
is limited to the subtree identified by extended all, or to the single
project identified by the simply extended project.

This change:
* removes the tree-level handling of project substitution, as the changes
  are local to a subtree.
* adjusts the View_Id to take into account the extends all projects.
* in Recursive_View, mark all subtree as extending the extends all View.
* modifies the aggregation of the sources at the extends all view level to
  include (potentially) all the objects of the extended closure.
* modify the algo to retrieve the actual objects/ali according to the new
  properties.

Change-Id: Icf64578c0ca87a86afa14961f10cc1f3b6f4d2f3
Depends-On: I40da660459a230fc76c970ded3b6bf570d5eea6e
TN: U325-030