Improve the DAG.
This implements means to compute shortest paths from the DAG, and allows
reporting cycles found in it if any. This also adds a cache so that we
don't need to compute the topological sort each time we need it.

With the new cyclic dependency check from the DAG, the old one from
GPR2.Project.Definition has been removed.

As preliminary work for U325-030

Change-Id: I204d7cc867ddf95635abcea194c91cbbf21e4490