Improve GPRCONFIG_PREFIX evaluation
Use a more sophisticated way of evaluating GPRCONFIG_PREFIX
in configuration chunks. Take into account that gprtools may
not be on the PATH (like during gprconfig testing), also try to
select proper gprtools installation when a gprtool is spawned
with a path prefix (it may be not first one on the path, and we
want to select consistent prefix dir).


Change-Id: Ic6aed9f15a82f3277368f6135b8f51f776c25c73
TN: SA23-009