Refactor the way view attributes are computed.
The view's "Attribute" function now fully computes the attributes values
based on the attributes definitions in Project.Registry.Attribute.

The various subprograms that were used to retrieve individually those
values are now simplified and call Attribute.

As a result, retrieval of Package Values and Attributes need to be
performed via the View, so as a result GPR2.Project.Pack is now a private
package, and only contains the raw object definition of a Package.

The parser is also reworked to benefit from the attribute computation when
referencing attributes. As a result it now need to work on the view object
rather than trying to build first the lists of variables and attributes
before building the object.

In order for the parser to see the package definition, it has been also
renamed to GPR2.Project.Parser, and GPR2.Parser hierarchy is now removed.

Finally, the GPR2.Project.Attribute object has been improved to include
missing information to fully handle all situations. In particular the
notion of Aliased attributes is now added to it (attributes that have
different names, such as Spec and Specification).

Part of U813-023

Change-Id: Ife12e67035c45f2915050e6550eb7d83d98fede2
Depends-On: Ie49c13da3ef7dbfc5f9b51f863e46b9f0ff9c4d1