Optimise retrieval of a source's other part.
The other part calculation is quite heavy, and is usually done twice: once
to check for the other part existence, then to actually retrieve it.

Introduce then a new function that returns undefined when the other part
does not exist, and use it in critical places such as
gpr2.project.source.artifacts.

TN: U722-008
Change-Id: Ie7877b2408b2a2f788489ca52ffebbc4a6a3d99a