Perf: revamp source dependencies cache and handling.
This patch makes sure that Dependency_Maps.Map is always used by ref and
are not actually part of an object.

The nice side effect is that now the various Source_Info that are part of
a single compilation unit are now sharing the same dependency map.

Together with the simplification in handling cache, merging the lists, this
reduces overall time on a typical (big) project by 40%.

TN: U722-008
Change-Id: Idcecbc79a69b2a8240587cacd17b3113fa97b1d6