Change the data structure for dependencies from map to vector
Building the list of dependencies is much faster as we just need to append
values to the vector rather than finding the right slot each time a value
is added.

Moreover, we don't really need a map to reason on dependencies, as the
list is always fully iterated.

TN: U722-008
Change-Id: I70780b351babf7a3526b2d813e74ea0ae10d4c7f