Improve multi-unit sources support.
First this change makes sure that the whole libgpr2 is using Unit_Index
when refering to units within multi-units sources, and moves this
definition to gpr2.ads for convenience (it's manipulated a bit everywhere).

Next the multi-unit list (GPR2.Unit.List) is completely re-implemented to
take into account actual multi-units, and single unit sources. No_Index
applies to both sources without units and single-unit sources, as does the
gpr project formalism. As a result, the list of units stored in a view need
another container to store them.

Also, this change adds two structures: GPR2.Project.Source.Source_Part
and GPR2.Unit.Source_Unit_Identifier that add index to the source
representation (Project.Source.Object for the first one and simple path
for the second one).

Then GPR2.Project.Unit_Info is fixed to store the Source_Part information
instead of just the Source object. The rest of the change is mostly the
result of this initial change.

Some minor changes in the way gprls reports multiunit sources have been
implemented, so that the end user can know what unit is actually
involved in the report.

TN: UA06-010
Change-Id: I0d675ec60ac867f296d65f2254ca5f17374f98e4
Depends-On: Ia955cc2c124f72d23f3f70e2862c0f7d947d968f