Detect unknown languages at an earlier stage
If a language passed to configuration phase is not declared in the KB
neither in any of the real compilers nor as a language that does not
require a compiler, do not add it to iterative search and instead report it
right away. This significantly speeds up configuration phase when unknown
language is involved.

Change-Id: I1ae61ad3a3010d24570efe12ebf1cca2c586d77c
TN: UA13-030, U809-008