Fix access discriminant failures.
The reference types we used in our custom containers were in general
wrong, as they did not add a refcount to the referenced value
(so the value could be removed/replaced and thus the ref become
a dangling pointer).

Also some tree iterators were storing temporary containers in the wrong
place, making refcounting next to impossible.

TN: UA14-020
Change-Id: I96110a95ee1f260226d50d8db24b50b026e80cfd