libgpr2: types defined in imported projects are directly visible.
when importing a project that defines some type for variables, this type
is directly visible from the importing project and can be used via its
base identifier.

This "feature" is for compatibility purpose as users can't control the
actual type used in case of overloading from different imports. So we
issue a warning with a suggestion for the proper fix.

TN: UA15-002
Change-Id: I0be41a5d20a9ccf828e18ce572ffd2f4274e4449