Fix ali parsing issue with 5.04a1.
The issue was that the ALI parser was running Tree.Update_Sources when
he couldn't resolve a unit name from a source path name. As it was called
while iterating the source of a project, the second call to update_sources
failed with the "try to temper a locked container" error.

This second pass was needed because the runtime is not added before the
other views. This is now fixed.

This change also fixes a couple of minor remaining issues with the
Source.Set container and in particular the iterator.

TN: UA11-003
Change-Id: I449a6159b68ecbddb56b57eeee97a91e1c6539aa