Update_Source_Parse: amend the Sources list instead of replacing it.
The list of sources used to be fully rebuilt after the sources are parsed
which lead to deallocation of the initial list, and copy of (potentially
very large) values of the list.

This patch amends the list as we go through it instead, reducing
significantly the processing time for projects with large number of sources.

TN: U722-008
Change-Id: I594830cca1cf15b2cd04779a8a9ccde677492f8c