Fix wrong unit name from ali file
The unit name as reported by the ali file was overriden just because of a
mixed casing reason.

The proper fix is to always apply the same casing to those unit names,
and not try to complexify unit names retrieval.

In this context, we decided to store unit names as upper case, to fully
convey that we don't try to apply the casing as written in the sources.

TN: UC02-045
Change-Id: I5d4ea8fd1334476f66c02210d2edf36ebf040d4e
Depends-On: If5694ae0b2cc64dcdf4aa10007e0ab6caa95d80e