Fix some issues with implicit projects, and report errors from parser
In Tree.Load_Autoconf, and in Project.Configuration.Create, do not
ignore parser errors, and add them to the tree's messages.

Also fix final error message when the project is the implicit project
in gprls and gprclean.

In GPR2.Project.Configuration.Create, take care of the implicit project
when declaring Project_Path.

Discovered while working on V503-010

Change-Id: I4157328bd4cc293e330fa4a1438666b3c7c12171