Further refactor the way source dirs are retrieved
This change now uses the exact same function when retrieving the
source directories or when we lookup source files.

This ensures that the returned list of source dirs is always coherent with
the way the sources are actually looked up.

In particular this fixes an issue with previous implementation where
Ignored_Source_Dirs attribute was not taken into account.

TN: V209-021
Change-Id: I75cee40a2087896a0a1c4c27706b7aed4f988049