Add View_Builder API to create projects programmatically
This allows generation of simple root projects with some attributes
customized from the default project.

This change removes the old handling of implicit projects and use the
new API to generate the default project.

GPR2.Path_Name.Implicit_Project is thus removed and the special hacks there
are not needed anymore.

TN: V602-035
Change-Id: I8b92fdfe995b7c3e182ef42cb8df44df6b9f4950