Correction of GPRinspect when using an unknown switch
When using an unknown switch GPRinspect raises a USAGE_ERROR exception
which is caught and we try to print an error message from the project
tree.
But following V523-014 the project loading was moved to a subprocess
which means we were trying to print messages from an unloaded tree
raising a CONSTRAINT_ERROR with an access check failed.
Added a test verifying that an USAGE_ERROR exception is properly raised
when an unknown switch or unknown switch value are set.

TN: V621-035
Change-Id: Ica259bd5f15c2161c94020edb5d5743cbf542366