Remove source lists from the tree object.
This removes the unneeded lists of sources from the tree object, using
the view ones as reference.

This also simplifies the way overloaded sources are handled by reasoning
only at the view level: if a view is overload, treat all the sources as
overloaded (because they're either used in the context of the overloading
view or are actually overloaded by a new source).

TN: V623-004
Change-Id: I2581c1f6def8d525baea384958778bd2203c6b2c