Functional testing of Source_Dirs attribute
This test makes sure the source parsing of libgpr2 is working as
intended when using the Source_Dirs attributes.
Additionally verifies the efficiency of the concomitant use of
Excluded_Source_Dirs, Ignore_Source_Sub_Dirs and Source_Dirs.
Improvement of the test driver to take into account attributes
that have a list as a value.
Correction of the regression of shared library suffix and prefix
test on windows due to its migration on the test driver

TN: V509-025
Change-Id: I2e0f3af22a01ea08f3289d7fc82640c08a0f49cd