Prototype of build database
This first version is completely separated from the current way of loading
sources, compilation unit information and dependencies calculation.

This allows for an incremental improvement of the build db while not
introducing right away regressions in the current tools.

Supported for now: loading sources and deducing compilation units from the
naming schema + Ada source parser.

Not tested for now: incremental update of the build database.

Not implemented: storage of the build db on disk, dependencies.

TN: V228-058
Change-Id: Ib745c1e684ee7c0ba5ee2496a81825eff5c43043