gpr parser: differentiate empty and no value
We need to differentiate cases where we know we can't evaluate values of
a variable or attribute from cases where said attribute holds an empty
list of values.

Else we might simply incorrectly set the value of an attribute to a wrong
values kind, leading to unexpected exceptions.

TH: V822-011
Change-Id: I46beb1f284aa0718794fcaf198bafbc3f1a3a05a