Switching from GNAT.Regexp to GPR2.FNMatch
On certain situation, if we had a lot of files and a lot of
attributes with filenames as index, retrieving the attribute
value was slowing down the process (GPRBuild and compilation
time for instance).
Switching from GNAT.Regexp to a more specific implementation
that suit our needs drastically reduced this behavior.

This change will be moved to GNATCOLL.FNMatch in the future.

TN: T915-017
Depends-On: Ie1fabaccdaece97fc2d18a913172c19027a18409
Change-Id: I4a8d7b7aaa00a0dd042c90354d7768c054c6bd06