Improve handling of root of subtrees.
The namespace root of a view is the project root of a compilable subtree
(so either the root project of the tree, or the projects aggregated by
an aggregate project). In case of aggregate projects, a view can have
several associated projects, in case it is shared between subtrees.

This change handles such situation, by changing the simple value by a list
of views.

TN: VA13-023
Change-Id: I9f047d7df53db0b833e467beec6999c1a23c1d51