Adding a test driver support for documentation and test generation
In depth documentation of the driver principle and how to properly
define a test scenario and testcases.
Addition of tools for the test driver such as creating, deleting and
running driver tests without launching the whole testsuite.
When creating a new driver test, generates a basic test skeleton to
adapt.

TN: VA25-007
Change-Id: Icdfa508065e3e8f026e875bbcb22e4e1ff04338d