testsuite_support/base_driver.py: Add aarch64-linux substitution rule
This change adds a substitution rule in the testsuite's framework
for aarch64-linux, similar to the rules that are already there for
other targets, such as x86_64-linux, x86-windows, etc. The names
of the platforms or triplets seem hardcoded, so this patch just
does the same.

This fixes 3 spurious failures when running the gpr2 testsuite
on native aarch64-linux configuration, with diffs like the following:

    | -Testing [...] --target (host)
    | +Testing [...] --target aarch64-linux

TN: VC16-033