Add some aarch64-linux-specific output files
The output of a couple of tests depend on how the knowledge base for
that platform has been defined, and in particular whether the KB
explicitly sets the value of some attributes instead of relying on
the default value when not set. This commit ajusts the expected output
of a couple of tests to match what's currently defined in the KB.

Note that a generic ".2" extension is chosen for the output file,
so as to facilitate its reuse for other platforms, in case new
platforms get tested in the future that needs the same output.

TN: W411-024