Better handling of runtime
The runtime view is now added to the tree after the second stage of the
tree parsing occurred: only after the context is set can we know that
the value of the Attribute "languages" is known.

This fix also brings better option to the end user: With_Runtime now only
controls the load of the runtime sources, but the runtime project will
still exist in case it is referenced by the parent view.

For eng/gpr/gpr-issues#34