gpr2next: fix "tree-source-dirs" test
While fixing this test, it was observed that calling `Update_Sources` on a
project without sources fails, because the tree database is undefined.
`Load_Autoconf` calls twice `Load` if there is no default configuration
file. `Tree_Db` is created only in the second load. The issue was that
if the project had no language available, as it is the case for an empty
project, then `Load_Autoconf` returns before calling the second `Load`.

The commit fixes this issue, and updates the test.

For eng/gpr/gpr-issues#67

Co-authored-by: Jerome Lambourg <lambourg@adacore.com>