API to obtain project external information
Provide an API to obtain information about typed and untyped externals.

A typed external is an external which has been assigned to a typed
variable. If several variables use the same external with a different
type, the types values union is returned as the possible values,
and the external is tagged as ambiguous.

Moreover, an external can be typed only if its type can be statically
determined, which means that:

```
type Version_Type is ...
var : Version_Type := external ("VERSION", "22.1") & "-optimized";
```
is not valid, in contrary to:

```
var : My_Type := external ("VAL1", external ("VAL2", "default-val2"));
```
which is valid, with VAL1 and VAL2 having the same type.

== Limitations ==

- External references are not processed during the external parsing.

Issue: gpr/gpr-issues/issues/82