Simplify dependencies lists generated by the Ada/Ali parsers
The dependency is changed from Source_Reference.Identifier to a list
of Name_Type, as there's no real need to keep the source location
information (only use is the warning
'<source>: with unit "foo" not found' where the actual location is not
really useful).

This allows simplifying the low level Ada parser, removing the need for
a Token Handler.

Ref eng/gpr/gpr-issues#81, eng/gpr/gpr-issues#113