gprinspect: Display configuration and tree messages in a dedicated "messages" field
Information, warnings and error messages related to the configuration and
tree loading are now available at the beginning of the output for both JSON
and text formats.

Until now, the errors were handled by `GPRtools.Options.Load_Project`
with `Handle_Errors` set. This function displayed logs and terminated the
program in case of error with the error code set. We cannot rely anymore
on this mechanism, as we need control over the output. GPRtools.Util.Fail_Program
is now used to reproduce this behavior.

Note that if gprinspect fails with the JSON display enabled, loading messages will be
displayed in the JSON, while gprinspect error will be in plain text after the JSON output.

For example:
```
{
  "messages": {
    "configuration": [
    ],
    "tree": [
      "lib.gpr:3:22: error: \"<path>/tmp/../obj-lib/\" cannot relocate object directory deeper than relocated build tree,",
      "lib.gpr:4:23: error: \"<path>/tmp/../lib/\" cannot relocate library directory deeper than relocated build tree,"
    ]
  }
}
gprinspect: unable to process project file prj/subdir/prj.gpr
```

Issue: eng/gpr/gpr-issues#197