Add check when relocated build tree goes out of root build dir
This checks that if the root project in the tree is not the topmost dir
then a relocated build won't try to access object dirs outside of the
relocated build tree.

This also fixes the profile of GPR2.Path_Name.Relative_Path in the process
to remove confusion between result.Name (that is the actual relative path)
and result.Value (that is an absolute path). Fix code here and there that
was using Relative_Path incorrectly.

For eng/gpr/gpr-issues#200