Fix Substitute_Variables_In_Configuration for "*" indexes
Until now, when "*" index was used in configuration, only the first
language in the languages list was checked. If the variable was not
found, then the empty string "" was returned, without error.

The expected behaviour is to check all languages, until we found
the variable for one of them. If the variable does not exist, an
exception is raised.

The message has also been modified to be in Messages and not in
Gnatcoll traces. Indeed, GPR2.KB.Configuration does not expect an
empty string, except if an error was raised, which is equivalent to
having an error message in Messages.

Warnings for each language are raised only in case of error.

Issue: eng/gpr/gpr-issues#155