Refine the build database DAG
In particular Artifacts manipulated as nodes are now better defined,
and while still not allowing a restore from disk allow to better
describe the real world artifacts we're manipulating.

The notion of actions is also better define and will allow a proper
population of the DAG as artifacts are added to it.

Build.Compilation_Input is now obsolete and is thus removed in favor
of children of Artifacts.

Ref eng/gpr/gpr-issues#151