gpr-kb: Test indexed variables in compiler_description
Indexed variable are not allowed in compiler_description. Test that
an error is well raised.

The `Messages` field of `All_Compilers` is not used properly. Errors
of parsing are raised in the embedded KB log messages (see
eng/gpr/gpr-issues#217 for more information). Thus, as the test checks
that the parsing fails, we need to check the KB embedded log messages.

Issue: eng/gpr/gpr-issues#155