Fix GPRinstall taking too long on uninstall mode
Went back on the logic of how we delete directory recursively in order
to only delete directory GPRinstall previously installed files in.
Introduced a cached system in order to lower FS calls to determine
whether or not a directory can be deleted. This enhance GPRinstall process
time on larger project.

Issue: eng/gpr/gpr-issues#251
Related-to: eng/gpr/gpr-issue#189