gpr2-kb: fix Parse_All_Dirs behavior for <directory> node with group 0
The <directory> node of <compiler_description> allows to compute a value
from a list of directories and files. If the field "group" is set to its
default value 0, then, according to the documentation, the value parsed
in the regex of <directory> shall be returned.

This was not the case until now. This commit fixes this feature.

Issue: eng/gpr/gpr-issues#155