doc: Use sphinx embedded search for offline usage
Pagefind does not work if the documentation HTML files are not hosted
by a web server (local, or not).

To ensure that the user can always have a search engine available, the
default sphinx search engine is used when the detected protocol is
neither `http` nor `https`.

To do so, the inherited RTD search template is extended with the
pagefind code. Both search engines have their own result `<div>`, so we
can display the correct one.

Co-Author: Boris Yakobowski <yakobowski@adacore.com>
Issue: eng/gpr/gpr-issues#260