Obtain dependencies name from ALI file
ALI files contain dependency information.

From the GCC documentation:
"This is a list of files, together with time stamp and checksum information.
 These are files on which the unit depends in the sense that recompilation
 is required if any of these units are modified."

Each dependency represents a build DB artifact that can be used an an
action input. Only the name is required, as timestamps and
checksums will be checked separately.

Issue: eng/gpr/gpr-issues#276