Use normalized paths to compute the view ID to avoid duplicate registrations
Let's take two project files, A and B. A imports B, and B imports A with
a limited with. If the path used to import a project contains symbolic
link as a directory, then this path was concatenated to the base directory
of the search. Because these symbolic links are not resolved and
because a view ID is computed with the path of a project file, an infinite
number of views were created.

So, for instance:

A/a.gpr: with "symlink-to-dir-B/b.gpr"
B/b.gpr: limited with "symlink-to-dir-A/a.gpr"

The following views were created during the project loading of A:
- View for A/a.gpr
- View for A/symlink-to-dir-B/b.gpr
- View for A/symlink-to-dir-B/symlink-to-dir-A/a.gpr
- View for A/symlink-to-dir-B/symlink-to-dir-A/symlink-to-dir-B/b.gpr
- ...

The fix to this issue is to always resolve the symbolic links when computing
the view ID. The view path remains however unresolved.